requirements.txt
praw==5.0.1
PyYAML==3.11
